requirements.txt
# Django & REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Environment & Configuration
python-dotenv>=1.0.0

# LLM & AI (Gemini - new SDK)
google-genai>=1.0.0

# Image Processing
Pillow>=10.0.0

# WhatsApp / Twilio
twilio>=8.0.0

# Utilities
pydantic>=2.5.3
httpx>=0.26.0
python-dateutil>=2.8.2
